Change constructors to take WebDriver parameter.
All the POM constructors have been changed to take a WebDriver parameter (which
is then immediately cast to a Browser instance). This is to make sure that
PageFactory.initElements calls the correct constructor (via reflection).

The alternative was to make a PageFactory object that had the correct class
information to call the correct constructor via reflection, but copy and paste
seems unnecessary with a decent comment.