WP-4217 - Update Tests to only open one browser window.
Update the test to not extend browser anymore. It now only opens one
browser widnow. Remove the constructor and do everything in the
@BeforeClass method. Remove the IOExceptions from the POM files so
that there is no need to catch them in the test. If an exception is
thrown the test will fail, so no need to handle it gracefully in this
case. Deleted commented out code and unused imports.