WP-4224 - Minor updates with issues that I ran into on my machine.
Add a wait for the error messages to appear before asserting that they
are displayed. Leaving the assert so that it is easy to tell what the
test is asserting, although it isn't necessary.

Add a wait for page load before the first send keys. The first send
keys seemed to be messing up and would only send the a random amount
of letters before moving on to the next input box.

Uncommented the next button call.