WP-4217 - Clean up the old Jenkins file a little bit.
Clean up the old jenkins file. Remove the cd before the run, just run
the tests in the jenkins workspace for now. They don't generate a lot
of stuff so this should be ok for the time being. Add a new stage for
the test run that is seperate from the checkout. Remove the echos.